AC analysis, new library, some minor changes

+++ com/sun/electric/tool/autotracing/Automodelling.java
@@ -236,7 +236,6 @@ public class Automodelling {
 
         @SuppressWarnings("null")
         private void addModellingParametersToModelScheme(Cell autoCell) {
-            EditingPreferences ep = EditingPreferences.getInstance();
 
             Cell basicCell = Job.getUserInterface().getCurrentCell();
 
@@ -246,16 +245,16 @@ public class Automodelling {
             Iterator<NodeInst> itrNod = autoCell.getNodes();
             // there is a patch
             while (itrNod.hasNext()) {
-               
-                 NodeInst ni = itrNod.next();
+
+                NodeInst ni = itrNod.next();
                 if (ni.toString().contains("5400TP035_core{")) {
-                   paramNode_1 = ni;
+                    paramNode_1 = ni;
                 }
                 if (ni.toString().contains("5400TP035_core_ac")) {
                     paramNode_2 = ni;
                 }
             }
-            
+
             itrNod = basicCell.getNodes();
             while (itrNod.hasNext()) {
                 NodeInst ni = itrNod.next();
@@ -265,16 +264,19 @@ public class Automodelling {
                 }
             }
             assert paramNode_1 != null;
-            assert paramNode_2 != null;
             assert modelParamNode != null;
-            
-            if (((paramNode_1.toString().contains("5400TP035_core{"))&&(modelParamNode.toString().contains("5400TP035_core{")))) {
-                paramNode_2.kill(); 
+
+            if (paramNode_2 != null) {
+                if (((paramNode_1.toString().contains("5400TP035_core{")) && (modelParamNode.toString().contains("5400TP035_core{")))) {
+                    paramNode_2.kill();
+                    paramNode_1.copyVarsFrom(modelParamNode);
+                }
+                if (((paramNode_2.toString().contains("5400TP035_core_ac")) && (modelParamNode.toString().contains("5400TP035_core_ac")))) {
+                    paramNode_1.kill();
+                    paramNode_2.copyVarsFrom(modelParamNode);
+                }
+            } else {
                 paramNode_1.copyVarsFrom(modelParamNode);
-            } 
-            if (((paramNode_2.toString().contains("5400TP035_core_ac"))&&(modelParamNode.toString().contains("5400TP035_core_ac")))){
-                paramNode_1.kill();
-                paramNode_2.copyVarsFrom(modelParamNode);
             }
 
             /* while (itrNod.hasNext()) {
